# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: دکان گھر - Urdu E-commerce & Logistics API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      parameters:
        - in: query
          name: type
          schema:
            type: string
            enum: [product, vehicle]
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Products
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - in: query
          name: sessionId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/track/{trackingNumber}:
    get:
      operationId: trackOrder
      tags: [orders]
      summary: Track an order by tracking number
      parameters:
        - in: path
          name: trackingNumber
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # Vehicles
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List vehicle bookings
      parameters:
        - in: query
          name: sessionId
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking address/rent (admin live edit)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # Payment Methods
...
  /admin/verify-pin:
    post:
      operationId: verifyAdminPin
      tags: [admin]
      summary: Verify admin PIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyAdminPinBody"
      responses:
        "200":
          description: PIN verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResponse"
        "401":
          description: Invalid PIN
  /admin/reset-pin:
    post:
      operationId: resetAdminPin
      tags: [admin]
      summary: Reset admin PIN (sends email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetAdminPinBody"
      responses:
        "200":
          description: Reset initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameUrdu:
          type: string
        type:
          type: string
          enum: [product, vehicle]
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - nameUrdu
        - type
        - sortOrder
        - createdAt
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameUrdu:
          type: string
        type:
          type: string
          enum: [product, vehicle]
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - name
        - nameUrdu
        - type
    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        nameUrdu:
          type: string
        icon:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        productNameUrdu:
          type: string
        quantity:
          type: number
        unit:
          type: string
        price:
          type: number
      required:
        - productId
        - productName
        - productNameUrdu
        - quantity
        - unit
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        trackingNumber:
          type: string
        sessionId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        deliverySlot:
          type: string
          enum: ["07:00", "11:00", "16:00"]
        paymentMethod:
          type: string
          enum: [cod, online]
        subtotal:
          type: number
        deliveryCharge:
          type: number
        total:
          type: number
        totalWeight:
          type: number
        status:
          type: string
          enum: [pending, packed, on_the_way, delivered]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - trackingNumber
        - sessionId
        - customerName
        - customerPhone
        - address
        - deliverySlot
        - paymentMethod
        - subtotal
        - deliveryCharge
        - total
        - totalWeight
        - status
        - items
        - createdAt
        - updatedAt
    CreateOrderBody:
      type: object
      properties:
        sessionId:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        deliverySlot:
          type: string
          enum: ["07:00", "11:00", "16:00"]
        paymentMethod:
          type: string
          enum: [cod, online]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required:
        - sessionId
        - customerName
        - customerPhone
        - address
        - deliverySlot
        - paymentMethod
        - items
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, packed, on_the_way, delivered]
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        vehicleId:
          type: integer
        vehicleName:
          type: string
        vehicleNameUrdu:
          type: string
        vehicleType:
          type: string
        customerName:
          type: string
        customerPhone:
          type: string
        passengersDetail:
          type: string
        luggageDetail:
          type: string
        pickupAddress:
          type: ["string", "null"]
        dropoffAddress:
          type: ["string", "null"]
        adminAddress:
          type: ["string", "null"]
        agreedRent:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - vehicleId
        - vehicleName
        - vehicleNameUrdu
        - vehicleType
        - customerName
        - customerPhone
        - passengersDetail
        - luggageDetail
        - status
        - createdAt
        - updatedAt
    CreateBookingBody:
      type: object
      properties:
        sessionId:
          type: string
        vehicleId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        passengersDetail:
          type: string
        luggageDetail:
          type: string
        pickupAddress:
          type: ["string", "null"]
        dropoffAddress:
          type: ["string", "null"]
      required:
        - sessionId
        - vehicleId
        - customerName
        - customerPhone
        - passengersDetail
        - luggageDetail
    UpdateBookingBody:
      type: object
      properties:
        adminAddress:
          type: ["string", "null"]
        agreedRent:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled]
...
    VerifyAdminPinBody:
      type: object
      properties:
        pin:
          type: string
      required:
        - pin
    AdminAuthResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
      required:
        - success
        - token
    ResetAdminPinBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email